+++ pom.xml
@@ -101,7 +101,7 @@
                 <version>0.0.8</version>
                 <configuration>
                     <launcher>launcher</launcher>
-                    <mainClass>at.fhhagenberg.sqe.App</mainClass>
+                    <mainClass>at.fhhagenberg.sqe.gui.App</mainClass>
                 </configuration>
             </plugin>
             <plugin>
