+++ src/main/java/at/fhhagenberg/sqe/backend/IElevator.java
@@ -1,4 +1,4 @@
-package at.fhhagenberg.sqe.model;
+package at.fhhagenberg.sqe.backend;
 
 /**
  * High-level interface to access and control the elevator system.
