+++ src/test/java/at/fhhagenberg/sqe/backend/ElevatorHardwareManagerTest.java
@@ -1,5 +1,6 @@
-package at.fhhagenberg.sqe.model;
+package at.fhhagenberg.sqe.backend;
 
+import at.fhhagenberg.sqe.model.Elevator;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
