+++ src/test/java/at/fhhagenberg/sqe/model/ElevatorModelFactoryTest.java
@@ -5,15 +5,19 @@ import static org.mockito.Mockito.when;
 
 import java.rmi.RemoteException;
 
+import at.fhhagenberg.sqe.backend.ElevatorHardwareManager;
+import at.fhhagenberg.sqe.backend.HardwareConnectionException;
+import at.fhhagenberg.sqe.backend.IElevatorManager;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
+import at.fhhagenberg.sqe.backend.IElevator;
 
 @ExtendWith(MockitoExtension.class)
 class ElevatorModelFactoryTest {
 	@Mock
-    IElevator iElevatorMock;
+	IElevator iElevatorMock;
 	
 	@Test
 	void testCreatingFactoryWithNoManagerFails() {
